application/single_app/requirements.txt
# requirements.txt
# All dependencies are pinned to exact versions for reproducible builds.
# To upgrade, update versions here and test before deploying.
# Generated/updated: 2026-04-07
pandas==2.2.3
azure-monitor-query==1.4.1
Flask==2.2.5
Flask-WTF==1.2.1
gunicorn==23.0.0
Werkzeug==3.1.5
requests==2.32.4
openai==1.98.0
docx2txt==0.8
Markdown==3.3.4
bleach==6.1.0
azure-cosmos==4.9.0
msal==1.31.0
Flask-Session==0.8.0
azure-ai-documentintelligence==1.0.2
numpy==1.26.4
scikit-learn==1.5.2
SciPy==1.15.1
joblib==1.4.2
threadpoolctl==3.5.0
azure-search-documents==11.5.3
python-dotenv==0.21.0
azure-ai-formrecognizer==3.3.3
pyjwt==2.9.0
markdown2==2.5.3
azure-mgmt-cognitiveservices==13.6.0
azure-identity==1.23.0
azure-ai-contentsafety==1.0.0
azure-storage-blob==12.24.1
azure-storage-queue==12.12.0
azure-keyvault-secrets==4.10.0
pypdf==6.4.0
python-docx==1.1.2
flask-executor==1.0.0
PyMuPDF==1.25.3
langchain-text-splitters==0.3.9
beautifulsoup4==4.13.3
openpyxl==3.1.5
xlrd==2.0.1
pillow==11.1.0
ffmpeg-binaries-compat==1.0.1
ffmpeg-python==0.2.0
semantic-kernel[mcp]==1.39.4
redis==5.2.1
pyodbc==5.2.0
PyMySQL==1.1.1
azure-monitor-opentelemetry==1.6.13
psycopg2-binary==2.9.10
Cython==3.0.12
pyyaml==6.0.2
aiohttp==3.13.3
html2text==2025.4.15
matplotlib==3.10.7
azure-cognitiveservices-speech==1.47.0
tiktoken==0.9.0
trafilatura==2.0.0
ultimate-sitemap-parser==1.5.0
PyGithub==2.6.0
simhash==2.1.2
networkx==3.4.2
graspologic==3.4.1
flask-limiter==3.12